+++ src/config.js
@@ -1,0 +1,4 @@
+const AUTHORIZATION = 'Basic nepeivinaGertruda';
+const END_POINT = 'https://17.ecmascript.pages.academy/cinemaddict';
+
+export { AUTHORIZATION, END_POINT };
