+++ src/main.js
@@ -2,14 +2,12 @@ import FilmsBoardPresenter from './presenter/films-board-presenter';
 import FilterPresenter from './presenter/filter-presenter.js';
 import FilmsModel from './model/films-model.js';
 import FilterModel from './model/filters-model';
-import FilmsApiService from './films-api-service.js';
-
-const AUTHORIZATION = 'Basic nepeivinaGertruda';
-const END_POINT = 'https://17.ecmascript.pages.academy/cinemaddict';
+import FilmsApiService from './api/films-api-service.js';
+import { AUTHORIZATION, END_POINT } from './config';
 
 const siteHeaderElement = document.querySelector('.header');
 const siteMainElement = document.querySelector('.main');
-const siteFooterlement = document.querySelector('.footer');
+const siteFooterElement = document.querySelector('.footer');
 
 const filmsModel = new FilmsModel(new FilmsApiService(END_POINT, AUTHORIZATION));
 const filterModel = new FilterModel();
@@ -20,7 +18,7 @@ const filmsBoardPresenter = new FilmsBoardPresenter(
   filmsModel,
   filterModel,
   siteHeaderElement,
-  siteFooterlement
+  siteFooterElement
 );
 
 filterPresenter.show();
