+++ src/model/comments-model.js
@@ -12,8 +12,7 @@ export default class CommentsModel extends Observable {
 
   init = async (film) => {
     try {
-      const comments = await this.#filmsApiService.getComments(film);
-      this.#comments = comments;
+      this.#comments = await this.#filmsApiService.getComments(film);
     } catch (err) {
       this.#comments = [];
     }
@@ -27,11 +26,10 @@ export default class CommentsModel extends Observable {
 
   addComment = async (updateType, update, updatedFilm) => {
     try {
-      const response = await this.#filmsApiService.addComment(update, updatedFilm);
-      const newComment = response;
+      const newComment = await this.#filmsApiService.addComment(update, updatedFilm);
       this.#comments = [newComment, ...this.#comments];
     } catch (err) {
-      throw new Error('Can\'t add comment');
+      throw new Error("Can't add comment");
     }
     this._notify(updateType, update, updatedFilm);
   };
@@ -40,7 +38,7 @@ export default class CommentsModel extends Observable {
     const index = this.#comments.findIndex((comment) => comment.id === update.id);
 
     if (index === -1) {
-      throw new Error('Can\'t delete unexisting comment');
+      throw new Error("Can't delete unexisting comment");
     }
 
     try {
@@ -48,7 +46,7 @@ export default class CommentsModel extends Observable {
       this.#comments = [...this.#comments.slice(0, index), ...this.#comments.slice(index + 1)];
       this._notify(updateType);
     } catch (err) {
-      throw new Error('Can\'t delete comment');
+      throw new Error("Can't delete comment");
     }
   };
 }
